Edit Completed

+++ ProductManager/src/main/java/net/codejava/product/dao/IProductDao.java
@@ -3,6 +3,6 @@ package net.codejava.product.dao;
 import net.codejava.product.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProductDao extends JpaRepository<Product, Long> {
+public interface IProductDao extends JpaRepository<Product, Integer> {
 
 }
